# docs/swagger.yaml
basePath: /api/v1
definitions:
  response.Response:
    properties:
      code:
        description: Code 0正确
        type: integer
      data: {}
      error:
        type: string
      msg:
        type: string
    type: object
  response.User:
    properties:
      avatar:
        description: 头像
        type: string
      created_at:
        description: 注册时间
        type: integer
      id:
        description: ID
        type: integer
      nickname:
        description: 昵称
        type: string
      status:
        description: 状态
        type: string
      user_name:
        description: 用户名
        type: string
    type: object
  response.UserResponse:
    properties:
      code:
        description: Code 0正确
        type: integer
      data:
        $ref: '#/definitions/response.User'
      error:
        type: string
      msg:
        type: string
    type: object
host: http://localhost:3000
info:
  contact: {}
  description: 这里写描述信息
  title: fastApi接口文档
  version: "1.0"
paths:
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口2
      parameters:
      - description: 密码
        in: query
        maxLength: 40
        minLength: 8
        name: password
        required: true
        type: string
      - description: 用户名
        in: query
        maxLength: 30
        minLength: 5
        name: user_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserResponse'
      summary: 用户登录接口1
      tags:
      - 用户相关接口
  /user/register:
    post:
      consumes:
      - application/json
      description: 用户注册接口
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 昵称
        in: query
        maxLength: 30
        minLength: 2
        name: nickname
        required: true
        type: string
      - description: 密码
        in: query
        maxLength: 40
        minLength: 8
        name: password
        required: true
        type: string
      - description: 确认密码
        in: query
        maxLength: 40
        minLength: 8
        name: password_confirm
        required: true
        type: string
      - description: 用户名
        in: query
        maxLength: 30
        minLength: 5
        name: user_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册接口
      tags:
      - 用户相关接口
swagger: "2.0"